Refactor create and delete functions
Simplify and clarify the onRequestPost() function in the create and delete functions. Instead of constructing an array of todo objects as JSON, both functions now return a JSON object with a 'status' key detailing what has been created or deleted. The 'Content-type' and 'Access-Control-Allow-Origin' headers have been added to the response object to permit communication with the API.